openurl.pl(process_line): Don't try to get "watch" setting if $type null
Was seeing errors like this one appearing in the status window every
time a URL was detected:

    Irssi: warning settings_get(openurl_watch_) : not found

For some reason `$type` was coming back as null. Didn't bother to figure
out why exactly, but everything else seems to be working okay, so
silence this error by ensuring `$type` has a value before using it to
query a potentially nonexistent setting.